@@ -21,10 +21,14 @@
 package com.netflix.spinnaker.cats.agent;
 
 public interface AgentIntervalAware {
-  /** @return Agent's interval to be scheduled at in milliseconds. */
+  /**
+   * @return Agent's interval to be scheduled at in milliseconds.
+   */
   Long getAgentInterval();
 
-  /** @return Agent's error interval to be scheduled at in milliseconds. */
+  /**
+   * @return Agent's error interval to be scheduled at in milliseconds.
+   */
   default Long getAgentErrorInterval() {
     return getAgentInterval();
   }
