@@ -6,11 +6,15 @@ import org.springframework.validation.Errors;
 /** Authorizes atomic operation description objects. */
 public interface DescriptionAuthorizer {
 
-  /** @param description The atomic operation description object this instance supports. */
+  /**
+   * @param description The atomic operation description object this instance supports.
+   */
   default boolean supports(Object description) {
     return true;
   }
 
-  /** @param description - The atomic operation description object. */
+  /**
+   * @param description - The atomic operation description object.
+   */
   void authorize(OperationDescription description, Errors errors);
 }
